Add tests for quadicons.
